requirements.txt
tensorflow-macos
streamlit
numpy
pandas
scikit-learn
matplotlib
h5py
protobuf